Separando as interfaces de tela no App.js em páginas separadas
